refactor: create infrastructure folder
