+++ app/src/main/java/com/example/partymaker/PartyApplication.java
@@ -8,7 +8,7 @@ import com.example.partymaker.data.api.NetworkManager;
 import com.example.partymaker.data.firebase.DBRef;
 import com.example.partymaker.data.repository.GroupRepository;
 import com.example.partymaker.data.repository.UserRepository;
-import com.example.partymaker.utils.system.MemoryManager;
+import com.example.partymaker.utils.infrastructure.system.MemoryManager;
 import com.example.partymaker.utils.ui.NotificationHelper;
 import com.google.firebase.FirebaseApp;
