+++ app/src/main/java/com/example/partymaker/data/api/NetworkManager.java
@@ -12,7 +12,7 @@ import androidx.annotation.NonNull;
 import androidx.lifecycle.LiveData;
 import androidx.lifecycle.MutableLiveData;
 import com.example.partymaker.utils.security.network.SSLPinningManager;
-import com.example.partymaker.utils.system.ThreadUtils;
+import com.example.partymaker.utils.infrastructure.system.ThreadUtils;
 import java.io.IOException;
 import java.net.HttpURLConnection;
 import java.net.URL;
