+++ app/src/main/java/com/example/partymaker/ui/chatbot/GptChatActivity.java
@@ -15,7 +15,7 @@ import com.example.partymaker.R;
 import com.example.partymaker.data.api.OpenAiApi;
 import com.example.partymaker.data.model.ChatMessageGpt;
 import com.example.partymaker.ui.adapters.ChatbotAdapter;
-import com.example.partymaker.utils.system.ThreadUtils;
+import com.example.partymaker.utils.infrastructure.system.ThreadUtils;
 import com.example.partymaker.viewmodel.GptViewModel;
 import com.google.android.material.appbar.MaterialToolbar;
 import java.io.IOException;
