+++ app/src/main/java/com/example/partymaker/ui/common/MainActivity.java
@@ -30,7 +30,7 @@ import com.example.partymaker.utils.auth.AuthenticationManager;
 import com.example.partymaker.utils.core.IntentExtrasManager;
 import com.example.partymaker.utils.core.ExtrasMetadata;
 import com.example.partymaker.utils.navigation.BottomNavigationHelper;
-import com.example.partymaker.utils.system.ThreadUtils;
+import com.example.partymaker.utils.infrastructure.system.ThreadUtils;
 import com.example.partymaker.utils.ui.LoadingStateManager;
 import com.example.partymaker.utils.ui.UiStateManager;
 import com.example.partymaker.utils.ui.UserFeedback;
