+++ app/src/main/java/com/example/partymaker/ui/group/AdminSettingsActivity.java
@@ -25,7 +25,7 @@ import com.example.partymaker.ui.common.MainActivity;
 import com.example.partymaker.utils.core.IntentExtrasManager;
 import com.example.partymaker.utils.core.ExtrasMetadata;
 import com.example.partymaker.utils.media.ImageCompressor;
-import com.example.partymaker.utils.system.ThreadUtils;
+import com.example.partymaker.utils.infrastructure.system.ThreadUtils;
 import com.google.firebase.database.DataSnapshot;
 import com.google.firebase.database.DatabaseError;
 import com.google.firebase.database.ValueEventListener;
