+++ app/src/main/java/com/example/partymaker/ui/group/ChatActivity.java
@@ -28,7 +28,7 @@ import com.example.partymaker.utils.core.IntentExtrasManager;
 import com.example.partymaker.utils.core.ExtrasMetadata;
 import com.example.partymaker.utils.security.encryption.GroupKeyManager;
 import com.example.partymaker.utils.security.encryption.GroupMessageEncryption;
-import com.example.partymaker.utils.system.ThreadUtils;
+import com.example.partymaker.utils.infrastructure.system.ThreadUtils;
 import com.example.partymaker.viewmodel.SimplifiedChatViewModel;
 import java.util.ArrayList;
 import java.util.Date;
