+++ app/src/main/java/com/example/partymaker/utils/auth/SecureAuthenticationManager.java
@@ -7,7 +7,7 @@ import com.example.partymaker.data.local.AppDatabase;
 import com.example.partymaker.data.repository.GroupRepository;
 import com.example.partymaker.data.repository.UserRepository;
 import com.example.partymaker.utils.core.AppConstants;
-import com.example.partymaker.utils.system.ThreadUtils;
+import com.example.partymaker.utils.infrastructure.system.ThreadUtils;
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
